Conflicts:
	DB/DBB/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBB0210011.raml
	DB/DBX/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbx/persistence/db/basic/DbT7056GappeiShichosonDac.java
	DB/DBX/standard/persistence/src/main/resources/jp/co/ndensan/reams/db/dbx/persistence/db/mapper/relate/kaigoserviceshurui/IKaigoServiceShuruiMapper.xml
	DB/DBX/standard/service/src/main/java/jp/co/ndensan/reams/db/dbx/service/core/kaigoserviceshurui/kaigoserviceshurui/KaigoServiceShuruiManager.java
	DB/DBX/standard/service/src/test/java/jp/co/ndensan/reams/db/dbx/service/core/kaigoserviceshurui/kaigoserviceshurui/KaigoServiceShuruiManagerTest.java
	DB/DBZ/standard/divcontroller/src/main/webapp/WEB-INF/standard/CommonChildDiv/chikushichosonselect/ChikuShichosonSelect.js
	DB/DBZ/standard/divcontroller/src/main/webapp/WEB-INF/standard/CommonChildDiv/kaigochohyoseigyokyotsu/KaigoChohyoSeigyoKyotsu.js
	DB/DBZ/standard/divcontroller/src/main/webapp/WEB-INF/standard/CommonChildDiv/kaigochohyoseigyokyotsu/KaigoChohyoSeigyoKyotsu.ts
	DB/DBZ/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/Test/TestChikuShichosonSelect.raml

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShichosonTokubetuKyufuService.java
@@ -39,7 +39,7 @@ extends ModelBase<ShichosonTokubetuKyufuServiceIdentifier,
      */
     public ShichosonTokubetuKyufuService(RString 市町村特別給付用サービスコード,
             FlexibleDate 市町村特別給付用サービス有効期間開始年月日,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         requireNonNull(市町村特別給付用サービスコード, UrSystemErrorMessages.値がnull.getReplacedMessage("市町村特別給付用サービスコード"));
         requireNonNull(市町村特別給付用サービス有効期間開始年月日, UrSystemErrorMessages.値がnull.getReplacedMessage("市町村特別給付用サービス有効期間開始年月日"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
@@ -106,7 +106,7 @@ extends ModelBase<ShichosonTokubetuKyufuServiceIdentifier,
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
