+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KogakuShikyuHanteiKekkaBuilderTest.java
@@ -31,7 +31,7 @@ public class KogakuShikyuHanteiKekkaBuilderTest extends DbcTestBase {
     private static HihokenshaNo 主キー名1;
     private static FlexibleYearMonth 主キー名2;
     private static HokenshaNo 主キー名3;
-    private static Decimal 主キー名4;
+    private static int 主キー名4;
 
     @BeforeClass
     public static void setUpClass() {
