+++ DB/DBC/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbc/persistence/db/basic/DbT3069KogakuGassanShinseishoKanyurekiDac.java
@@ -51,7 +51,7 @@ public class DbT3069KogakuGassanShinseishoKanyurekiDac implements ISaveable<DbT3
             HokenshaNo 保険者番号,
             RString 整理番号,
             RString 加入歴番号,
-            Decimal 履歴番号) throws NullPointerException {
+            int 履歴番号) throws NullPointerException {
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(対象年度, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年度"));
         requireNonNull(保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("保険者番号"));
